requirements.txt
accelerate==0.18.0
aes==1.2.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
apturl==0.5.2
async-timeout==4.0.2
attrs==19.3.0
bitsandbytes==0.38.1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
datasets==2.11.0
dbus-python==1.2.16
defer==1.0.6
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
Flask==2.2.2
Flask-Cors==3.0.10
flexgen==0.1.7
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gradio==3.25.0
gradio-client==0.1.3
h11==0.14.0
httpcore==0.17.0
httplib2==0.14.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==2.8
importlib-metadata==5.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linkify-it-py==2.0.0
lit==16.0.2
llama-cpp-python==0.1.36
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.10.4
networkx==3.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
orjson==3.8.10
packaging==23.1
pandas==1.5.1
peft==0.3.0.dev0
pexpect==4.6.0
Pillow==9.5.0
protobuf==3.6.1
psutil==5.9.5
PuLP==2.7.0
pyaes==1.6.1
pyarrow==11.0.0
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.7
pydub==0.25.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.5
pyxdg==0.26
PyYAML==5.3.1
regex==2023.3.23
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
responses==0.18.0
rsa==4.9
rwkv==0.7.3
safetensors==0.3.0
SecretStorage==2.3.1
semantic-version==2.10.0
sentencepiece==0.1.98
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
ssh-import-id==5.10
starlette==0.26.1
sympy==1.11.1
systemd-python==234
Telethon==1.26.0
texttable==1.6.2
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
uvicorn==0.21.1
wadllib==1.3.3
websocket-client==0.53.0
websockets==11.0.2
Werkzeug==2.2.2
xkit==0.0.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.11.0
